Move notebook converter script into github utils
