+++ .github/utils/code_and_docs.sh
@@ -1,10 +1,6 @@
 # Apply Black
 black .
 
-# Convert tutorial notebooks into webpages
-cd docs/_src/tutorials/tutorials/
-python3 convert_ipynb.py
-
 # Generate the API documentation
 set -e   # Fails on any error in the following loop
 cd docs/_src/api/api/
@@ -12,6 +8,10 @@ for file in ../pydoc/* ; do
     echo "Processing" $file
     pydoc-markdown "$file"
 done
+cd ../../../../
+
+# Convert tutorial notebooks into webpages
+python convert_notebooks_into_webpages.py
 
 # Generate OpenAPI docs
 python .github/utils/generate_openapi_specs.py
